+++ doc/changelog.rst
@@ -3,6 +3,7 @@
 Release Notes
 =============
 
+- :feature:`-` Instead of setting a flag somewhere, pretalx now has an explicit "go live" button. There is also a corresponding signal plugins can listen to (please refer to the plugin documentation for further information).
 - :bug:`463` Don't show a 404 error if a reviewer tries to review their own submission, but show an error message instead.
 - :feature:`-` For organisers, the submission/talk API endpoints now contain the question answers given by the speakers.
 - :feature:`-` Schedules now contain a search bar to filter talks by title or speaker.